Move eviction types into eviction file
